requirements.txt
python-dotenv>=1.0.0,<2
fastapi>=0.115.0,<1
uvicorn[standard]>=0.32.0,<1
jinja2>=3.1.0,<4
python-multipart>=0.0.9,<1
requests>=2.32.0,<3
